Restructure object access, and finish bootloader protocol
Generated objects are now using a set of sub objects, which all implement
`SubObjectAccess`. These include:
- ScalarField<T> for basic types, like u8 or i32.
- ByteField<N> for storing OctetStrings of size N.
- NullTermByteField<N> for storing visible and unicode strings. These will
  automatically append null terminators for write values smaller than the
  buffer.

Any object can implement the `ProvidesSubObjects` trait, which maps sub
index to metadata and the attribute which implements the subobject, and it
will automatically impl the `ObjectRawAccess` trait. This makes it easier
to create objects which impl `ObjectRawAccess`, and I think it should
decrease code size, because some of the code that was generated for each
object type can be shared by all objects.

Additonally, the CallbackObject now takes a `&dyn ObjectRawAccess` instead
of individual functions, which seems better all around.